refactor(advisory-convergence): extract review-clause into shared module
Move the latest-review Clause 1 evidence (types, the pure
resolveLatestCopilotReviewClause evaluator, the fetchReviewsAndHeadCommit
GraphQL fetch, and isVerifiedCopilotAuthor) out of advisory-convergence.mts
into a new small src/scripts/review-clause.mts module, and move the shared
ghGraphql wrapper into gh-exec.mts alongside ghText/ghApiJson.

No behavior change: advisory-convergence.mts re-imports every moved symbol
verbatim and keeps re-exporting AdvisoryConvergenceReviewClause for any
existing consumer. This groundwork lets a second, independent caller reuse
the exact same review-coverage evidence the real idd-advisory-convergence
gate uses, instead of a second ad-hoc GraphQL path, without pulling in
advisory-convergence.mts's full claim/waiver/disposition machinery.

Refs #1806
